Move CivilopediaImageGetters object to own file
